# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server Petstore server.",
        "title": "Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:8085",
    "paths": {
        "/gest": {
            "get": {
                "description": "Ritorna tutti gli ordini contenuti nella Collection",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ordini"
                ],
                "summary": "List All Ordini",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "description": "Crea un record nella Collection Ordini",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ordini"
                ],
                "summary": "Post one Ordine",
                "parameters": [
                    {
                        "description": "Inserisci un Ordine",
                        "name": "json",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/gest/{numeroOrdine}": {
            "get": {
                "description": "GET un record nella Collection Ordini",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ordini"
                ],
                "summary": "GET one Ordine",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Numero Ordine dell'Ordine",
                        "name": "numeroOrdine",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "put": {
                "description": "Esegui l'update di un Ordine",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ordini"
                ],
                "summary": "Update one Ordine",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Numero Ordine dell'Ordine",
                        "name": "numeroOrdine",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Modifica un Ordine",
                        "name": "json",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "description": "Esegui il delete di un Ordine",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ordini"
                ],
                "summary": "Delete one Ordine",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Numero Ordine dell'Ordine",
                        "name": "numeroOrdine",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    }
}